refactor: Refactoring some methods
Refactoring some methods
